[POAE7-756]split file at java level

+++ oap-ape/ape-native/src/com_intel_ape_ParquetReaderJNI.h
@@ -27,10 +27,10 @@ extern "C" {
 /*
  * Class:     com_intel_ape_ParquetReaderJNI
  * Method:    init
- * Signature: (Ljava/lang/String;Ljava/lang/String;ILjava/lang/String;JJ)J
+ * Signature: (Ljava/lang/String;Ljava/lang/String;ILjava/lang/String;II)J
  */
 JNIEXPORT jlong JNICALL Java_com_intel_ape_ParquetReaderJNI_init
-  (JNIEnv *, jclass, jstring, jstring, jint, jstring, jlong, jlong);
+  (JNIEnv *, jclass, jstring, jstring, jint, jstring, jint, jint);
 
 /*
  * Class:     com_intel_ape_ParquetReaderJNI
